+++ integrate_test/helloworld/tests/integration/helloworld_test.go
@@ -23,12 +23,46 @@ import (
 )
 
 import (
+	tripleConstant "github.com/dubbogo/triple/pkg/common/constant"
+
 	"github.com/stretchr/testify/assert"
 )
 
 import (
 	dubbo3pb "github.com/apache/dubbo-go-samples/api"
 )
+
+func TestStreamSayHello(t *testing.T) {
+	ctx := context.Background()
+	ctx = context.WithValue(ctx, tripleConstant.TripleCtxKey(tripleConstant.TripleRequestID), "triple-request-id-demo")
+	req := dubbo3pb.HelloRequest{
+		Name: "laurence",
+	}
+
+	r, err := greeterProvider.SayHelloStream(ctx)
+	assert.Nil(t, err)
+
+	for i := 0; i < 2; i++ {
+		err := r.Send(&req)
+		assert.Nil(t, err)
+	}
+
+	rspUser := &dubbo3pb.User{}
+	err = r.RecvMsg(rspUser)
+	assert.Nil(t, err)
+	assert.Equal(t, "hello laurence", rspUser.Name)
+	assert.Equal(t, "123456789", rspUser.Id)
+	assert.Equal(t, int32(18), rspUser.Age)
+
+	err = r.Send(&req)
+	assert.Nil(t, err)
+
+	err = r.RecvMsg(rspUser)
+	assert.Nil(t, err)
+	assert.Equal(t, "hello laurence", rspUser.Name)
+	assert.Equal(t, "123456789", rspUser.Id)
+	assert.Equal(t, int32(19), rspUser.Age)
+}
 
 func TestSayHello(t *testing.T) {
 	req := &dubbo3pb.HelloRequest{
